ar71xx: add support for OCEDO Ursus
This commit adds support for the OCEDO Ursus

SOC:    Qualcomm QCA9558 (Scorpion)
RAM:    128MB
FLASH:  16MiB
WLAN1:  QCA9558 2.4 GHz 802.11bgn 3x3
WLAN2:  QCA9880 5 GHz 802.11nac 3x3
INPUT:  NONE :(
LED:    LAN 1, LAN 2, WiFi 2.4, WiFi 5
Serial: Header in the middle of the to side of board
        Pinout is 3.3V - GND - TX - RX (Arrow Pad is 3.3V)
        The Serial setting is 115200-8-N-1.

Tested and working:
 - Ethernet
 - 2.4 GHz WiFi
 - 5 GHz WiFi
 - TFTP boot from ramdisk image
 - Installation via ramdisk image
 - OpenWRT sysupgrade
 - Buttons
 - LEDs

Installation seems to be possible only through booting an OpenWRT
ramdisk image.

Connect the serial. Wwhen asked hit any key to stop autoboot. It will load a
ramdisk image named 'ursus-uImage-initramfs-lzma.bin' from 192.168.100.8.

Note: To connect the serial, the device has to be opend. The metal cover
is not only screwed, but also glued to the casing. You might want to
replace the glue, after the flashing process, to enshure weather resistance.

Note: depending on the present software, the device might also try to
pull a file called 'ursus-uimage-factory'. Only the name differs, it
is still used as a ramdisk image.

Wait for the ramdisk image to boot. OpenWRT can be written to the flash
via sysupgrade or mtd.

Due to the flip-flop bootloader which we not (yet) support, you need to
set the partition the bootloader is selecting. It is possible from the
initramfs image with

 > fw_setenv bootcmd run bootcmd_1

Afterwards you can reboot the device.